+++ lang/zh_CN/Audio/ExpansionPackageLines.lua
@@ -148,15 +148,6 @@ return {
     ['$LuaShanxi1'] = '连舫逾万艘，带甲千万人',
     ['$LuaShanxi2'] = '率彼东南路，将定一举勋',
     ['~ExWangcan'] = '虽无铅刀用，庶几奋薄身……',
-    -- 周处
-    ['$LuaXianghai1'] = '快快闪开，伤到你们可就不好啦，哈哈哈哈哈哈！',
-    ['$LuaXianghai2'] = '你自己撞上来的，这可怪不得小爷我！',
-    ['$LuaChuhai1'] = '有我在此，安敢为害？！',
-    ['$LuaChuhai2'] = '小小孽畜，还不伏诛？！',
-    ['$LuaChuhai3'] = '此番不成，明日再战！', -- 失败
-    ['$LuaZhangming1'] = '心怀远志，何愁声名不彰！',
-    ['$LuaZhangming2'] = '从今始学，成为有用之才！',
-    ['~ExZhouchu'] = '改励自砥，誓除三害……',
     -- 界孙策
     ['$LuaJiang1'] = '我会把胜利带回江东！',
     ['$LuaJiang2'] = '天下英雄，谁能与我一战？',
@@ -317,14 +308,6 @@ return {
     ['$LuaDinghan1'] = '益国之事，虽死弗避',
     ['$LuaDinghan2'] = '杀身有地，报国有时！',
     ['~ExShenXunyu'] = '宁鸣而死，不默而生',
-    -- 神孙策
-    ['$LuaYingba1'] = '从我者可免，拒我者难容！',
-    ['$LuaYingba2'] = '卧榻之侧，岂容他人鼾睡！',
-    ['$LuaFuhai1'] = '翻江复倒海，六合定乾坤！',
-    ['$LuaFuhai2'] = '力攻平江东，威名扬天下！',
-    ['$LuaPinghe1'] = '不过胆小鼠辈，吾等有何惧哉！',
-    ['$LuaPinghe2'] = '只可得胜而返，岂能败战而归！',
-    ['~ExShenSunce'] = '无耻小人！胆敢暗算于我……',
     -- 文鸯
     ['$LuaQuedi1'] = '力摧敌阵，如视天光破云！',
     ['$LuaQuedi2'] = '让尔等有命追，无命回！',
@@ -335,15 +318,6 @@ return {
     ['$LuaChoujue1'] = '血海深仇，便在今日来报！',
     ['$LuaChoujue2'] = '取汝之头，以祭先父！',
     ['~ExWenyang'] = '半生功业，而见疑于一家之言，岂能无怨！',
-    -- 神太史慈
-    ['$LuaDulie1'] = '素来言出必践，成吾信义昭彰！',
-    ['$LuaDulie2'] = '小信如若不成，大信将以何立？',
-    ['$LuaPowei1'] = '君且城中等候，待吾探敌虚实！', -- 一般破围语音
-    ['$LuaPowei2'] = '弓马骑射洒热血，突破重围显英豪！', -- 破围成功
-    ['$LuaPowei3'] = '敌军尚犹严防，有待明日再看！', -- 破围失败
-    ['$LuaShenzhu1'] = '力引强弓百斤，矢出贯手著棼！',
-    ['$LuaShenzhu2'] = '箭既已在弦上，吾又岂能不发！',
-    ['~ExShenTaishici'] = '魂归……天地……',
     -- 曹金玉-十周年
     ['$LuaYuqi1'] = '孤影独泣，困于隅角',
     ['$LuaYuqi2'] = '向隅而泣，黯然伤感',
