+++ lang/zh_CN/Audio/LayingPlansWisdomPackageLines.lua
@@ -1,0 +1,30 @@
+-- lines for LayingPlans Wisdom Package
+
+return {
+    -- 周处
+    ['$LuaXianghai1'] = '快快闪开，伤到你们可就不好啦，哈哈哈哈哈哈！',
+    ['$LuaXianghai2'] = '你自己撞上来的，这可怪不得小爷我！',
+    ['$LuaChuhai1'] = '有我在此，安敢为害？！',
+    ['$LuaChuhai2'] = '小小孽畜，还不伏诛？！',
+    ['$LuaChuhai3'] = '此番不成，明日再战！', -- 失败
+    ['$LuaZhangming1'] = '心怀远志，何愁声名不彰！',
+    ['$LuaZhangming2'] = '从今始学，成为有用之才！',
+    ['~ExZhouchu'] = '改励自砥，誓除三害……',
+    -- 神太史慈
+    ['$LuaDulie1'] = '素来言出必践，成吾信义昭彰！',
+    ['$LuaDulie2'] = '小信如若不成，大信将以何立？',
+    ['$LuaPowei1'] = '君且城中等候，待吾探敌虚实！', -- 一般破围语音
+    ['$LuaPowei2'] = '弓马骑射洒热血，突破重围显英豪！', -- 破围成功
+    ['$LuaPowei3'] = '敌军尚犹严防，有待明日再看！', -- 破围失败
+    ['$LuaShenzhu1'] = '力引强弓百斤，矢出贯手著棼！',
+    ['$LuaShenzhu2'] = '箭既已在弦上，吾又岂能不发！',
+    ['~ExShenTaishici'] = '魂归……天地……',
+    -- 神孙策
+    ['$LuaYingba1'] = '从我者可免，拒我者难容！',
+    ['$LuaYingba2'] = '卧榻之侧，岂容他人鼾睡！',
+    ['$LuaFuhai1'] = '翻江复倒海，六合定乾坤！',
+    ['$LuaFuhai2'] = '力攻平江东，威名扬天下！',
+    ['$LuaPinghe1'] = '不过胆小鼠辈，吾等有何惧哉！',
+    ['$LuaPinghe2'] = '只可得胜而返，岂能败战而归！',
+    ['~ExShenSunce'] = '无耻小人！胆敢暗算于我……'
+}
